update på account Leaderboard.kt

+++ app/src/main/java/com/example/unomemo/leaderboard/Leaderboard.kt
@@ -1,7 +1,10 @@
 package com.example.unomemo.leaderboard
 
+import com.google.firebase.firestore.Blob
+
+
 data class Leaderboard(
     var uid: String="",
     var navn: String ="",
-    var poengsum: Int = -1
+    var poengsum: Long =-1
 )
